sherlock_back/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
bcrypt==3.2.0
bidict==0.21.2; python_version >= '3.6'
cffi==1.14.4
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
flake8-flask==0.9.3
flake8==3.8.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
flask-cors==3.0.9
flask-httpauth==4.2.0
flask-socketio==5.0.1
flask-sqlalchemy==2.4.4
flask==1.1.2
flask-restplus==0.13.0
flask-restful-swagger==0.20.2
itsdangerous==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
jinja2==2.11.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
marshmallow-enum==1.5.1
marshmallow==3.10.0
mccabe==0.6.1
pycodestyle==2.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyflakes==2.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pymysql==0.10.1
python-engineio==4.0.0
python-socketio==5.0.4
python-dotenv==0.15.0
pyyaml==5.3.1
r2c-py-ast==0.1.0b1; python_version >= '3.6' and python_version < '4.0'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy-utils==0.36.8
sqlalchemy==1.3.22
unidecode==1.1.2
werkzeug==1.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
